# Faster popcnt_neon_vcnt()