Old code (GCC 7.4.0):

```
./speed_arm_g++ 1000000 100
lookup-8                  ... time = 0.088076 s
lookup-64                 ... time = 0.088084 s (speedup: 1.00)
bit-parallel              ... time = 0.067484 s (speedup: 1.31)
bit-parallel-optimized    ... time = 0.046986 s (speedup: 1.87)
bit-parallel-mul          ... time = 0.034938 s (speedup: 2.52)
bit-parallel32            ... time = 0.112424 s (speedup: 0.78)
bit-parallel-optimized32  ... time = 0.086487 s (speedup: 1.02)
harley-seal               ... time = 0.015675 s (speedup: 5.62)
builtin-popcnt            ... time = 0.032368 s (speedup: 2.72)
builtin-popcnt32          ... time = 0.065658 s (speedup: 1.34)
builtin-popcnt-unrolled   ... time = 0.023990 s (speedup: 3.67)
builtin-popcnt-unrolled32 ... time = 0.048273 s (speedup: 1.82)
neon-vcnt                 ... time = 0.011393 s (speedup: 7.73)
neon-HS                   ... time = 0.016071 s (speedup: 5.48)
```

New code (GCC 7.4.0):

```
./speed_arm_g++ 1000000 100
lookup-8                  ... time = 0.087958 s
lookup-64                 ... time = 0.087984 s (speedup: 1.00)
bit-parallel              ... time = 0.067376 s (speedup: 1.31)
bit-parallel-optimized    ... time = 0.046874 s (speedup: 1.88)
bit-parallel-mul          ... time = 0.034778 s (speedup: 2.53)
bit-parallel32            ... time = 0.112359 s (speedup: 0.78)
bit-parallel-optimized32  ... time = 0.086439 s (speedup: 1.02)
harley-seal               ... time = 0.015356 s (speedup: 5.73)
builtin-popcnt            ... time = 0.032613 s (speedup: 2.70)
builtin-popcnt32          ... time = 0.065623 s (speedup: 1.34)
builtin-popcnt-unrolled   ... time = 0.023659 s (speedup: 3.72)
builtin-popcnt-unrolled32 ... time = 0.048044 s (speedup: 1.83)
neon-vcnt                 ... time = 0.008637 s (speedup: 10.18)
neon-HS                   ... time = 0.015856 s (speedup: 5.55)
```