@WojciechMula 2019-12-31
Kim, thanks a lot, that's really cool! What machine do you have? What CPU?

(My Raspberry died and another machine I had access to also died, so no more real-world-ARM tackling)