@kimwalisch 2020-01-01
> Kim, thanks a lot, that's really cool! What machine do you have? What CPU?

The benchmark was run on a a1.xlarge instance (ARM 64-bit, 4 vCPUs, 2.3 GHz, AWS Graviton Processor, 8 GiB memory) in the Amazon EC2 cloud. Note that the code did not compile out of the box, I had to remove ```-mfpu=neon``` from the Makefile because that flag does not seem to exist on ARM64.